refactor(session): split ReaderSession by what each method is for
`session.rs` was 1,630 lines: one `impl ReaderSession` with 102 methods, running from opening a PDF
to lassoing ink to writing an annotated export. Opening a file to change how crop margins work meant
loading the whole reading engine, and every diff to the session read as a diff to all of it.

Rust allows several `impl ReaderSession` blocks across a crate, and a private field stays visible to
a child module, so this is a move of method bodies with no change to the type:

  mod.rs      318   the struct, `Gesture`, lifecycle and the plain accessors
  open.rs     264   one open_* per format and its _with_store twin, meeting at `assemble`
  render.rs   241   the render path, the fit/crop map, the page cache and prefetch
  view.rs     279   display settings and the reflow typography that triggers repagination
  navigate.rs 165   gestures, jumps, outline, links, and the text-under-a-point queries
  ink.rs      299   the stroke model, autosave, and the sidecar flush/retry
  select.rs   135   lasso selection, the ink clipboard, and PDF export

Seven private helpers are now `pub(super)` because a sibling calls them —
`invalidate_render_cache`, `cached_crop_bbox`, `expand_crop`, `flush_ink`, `autosave_ink`,
`load_ink_for_current_page`, `load_layer_for_page`. `pub(super)` here means visible within
`crate::session` and no wider, so nothing reaches the crate's public surface. That those seven are
the entire cross-file surface is itself the argument for these boundaries: everything else a group
needs, it already owned.

`view.rs` puts display settings and typography together on purpose. They look like two concerns and
are one: both are the Adjust sheet, and the interesting property is which of them merely repaints
and which repaginates and so has to hold the reader on the same words.

Verified as a move, not a rewrite: all 102 method bodies are byte-identical to master except two,
and both of those differ only by the `default_supernote` -> `default_tablet_epd` rename and the
vendor-neutral comment from the IR-7 commit earlier on this branch.

`session_tests.rs` moved into the module directory beside the code it tests; its one
`include_bytes!` of the shared EPUB fixture needed the extra `..`. All gates green — 747 host
tests, 166 Kotlin, and the aarch64-android cross-check.
